Refactoring
-Rename de clases y reestructuración de carpetas.
-Dotenv variables de pool de DB
