+++ enviroments/dev.env
@@ -1,1 +1,6 @@
-PORT = 3001
+APPLICATION_PORT = 3001
+DATABASE_USER = postgres
+DATABASE_PASSWORD = postgres
+DATABASE_HOST = localhost
+DATABASE_NAME = access_db
+DATABASE_PORT = 5432
