+++ enviroments/pro.env
@@ -1,1 +1,6 @@
-PORT = 3000
+APPLICATION_PORT = 3001
+DATABASE_USER = 
+DATABASE_HOST = 
+DATABASE_PASSWORD = 
+DATABASE_NAME = 
+DATABASE_PORT =
