+++ src/main/Configuration.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
+import LoginRepositoryPostgreSQL from "./login/infraestructure/repositories/LoginRepositoryPostgreSQL";
 import { container } from "tsyringe";
-import LoginRepositoryPostgreSQL from "./login/infraestructure/implementations/LoginRepositoryPostgreSQL";
 
 container.register("LoginRepository", {
     useClass: LoginRepositoryPostgreSQL
