--- src/main/index.ts
@@ -1,5 +1,0 @@
-import 'reflect-metadata';
-import app from './app';
-
-app.listen(app.get('port'));
-console.log(`Listening on http://localhost:${app.get('port')}`);
