+++ src/main/login/aplication/adapters/LoginAdapter.ts
@@ -1,7 +1,7 @@
-import { autoInjectable, injectable } from "tsyringe";
-import { HandledError, isError } from "../../../shared/models/HandledError";
-import { User } from "../../domain/models/User";
 import LoginUseCase from "../../domain/services/LoginUseCase";
+import { injectable } from "tsyringe";
+import { HandledError, isError } from "../../../shared/domain/models/HandledError";
+import { User } from "../../domain/models/User";
 import { LoginResponseBodyView } from "../models/LoginResponseBodyView";
 
 @injectable()
