+++ src/main/shared/domain/models/HandledError.ts
