+++ src/main/shared/domain/services/PostgreSqlConnectionPool.ts
@@ -1,0 +1,9 @@
+import { Pool } from 'pg';
+
+export const PostgreSqlConnectionPool = new Pool({
+    user: process.env.DATABASE_USER,
+    host: process.env.DATABASE_HOST,
+    password: process.env.DATABASE_PASSWORD,
+    database: process.env.DATABASE_NAME,
+    port: Number(process.env.DATABASE_PORT)
+});
